file_list 调整结构
