+++ internal/app/commands/notify.go
@@ -1,0 +1,4 @@
+package commands
+
+// 通知
+var Notify func(cmd any)
